Adding the click listener to the recycler view
Humph. Adding various bits of data binding. There were a number of issues here adapting "simple" code from other places. A big complication was making the binding of the clicklistener to the view take place in the ViewHolder rather than onBind as apparently this is more performant as it's done once rather than every time a view gets bound. Marginal I'd have said but if it's supposed to be better practise then it's probably worth practising.