Adding navigation from the RecyclerView to a specific brewery.
Creating the navigation graph.
Passing the arguments from the ListFragment.
Making the Brewery Data object parcelable so it can be acepted as a custom argument.
Opening the info activity.
Creating a Viewmodel to hold the data.
Making the MainActivity launch with singletop so I return to the List rather than the Map.