Updating the Action Bar's functionality
So there were toolbars in each layout. To align the app with the navigation graph it's better to have a single toolbar in the Main Activity which runs from the navigation graph.
This is customised where appropriate with a custom back arrow and toolbar title. The ability to use the up arrow is also added.