Add moshi to the project
This is to allow me to pull the brewery information from the geojson and convert it from a GeoJson Feature into a Brewery object.
Adding navigation so I can go from the map to a selected brewery.
There were also some dependency updates when I added moshi to the build.gradle